pyproject.toml
[tool.poetry]
name = "django-utm-tracker"
version = "0.5.1"
description = "Django app for extracting and storing UTM tracking values."
license = "MIT"
authors = ["YunoJuno <code@yunojuno.com>"]
maintainers = ["YunoJuno <code@yunojuno.com>"]
readme = "README.md"
homepage = "https://github.com/yunojuno/django-utm-tracker"
repository = "https://github.com/yunojuno/django-utm-tracker"
documentation = "https://github.com/yunojuno/django-utm-tracker"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 2.2",
    "Framework :: Django :: 3.0",
    "Framework :: Django :: 3.1",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
]
packages = [{include="utm_tracker"}]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2 || ^3.0"
django-request-log = "^0.0.12"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-cov = "*"
pytest-django = "*"
coverage = "*"
pylint = "*"
flake8 = "*"
flake8-bandit = "*"
flake8-docstrings = "*"
isort = "*"
mypy = "*"
pre-commit = "*"
black = {version = "*", allow-prereleases = true}
freezegun = "*"
tox = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
